+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 **Features**
 - Added an API
 - Added Storage Drawers integration
+- Added handling for patterns that return the same item
 - Increased cable recipe to 12 cables
 
 ### 0.7.8